4.flask_projects/7.DBSTUDY/2.features/requirements.txt
Flask==2.3.2
python-dotenv==1.0.1
